Refactor feedback and grading page
