+++ frontend/src/app/layout.tsx
@@ -18,13 +18,13 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en" className="bg-white">
-      <body className={inter.className + " min-h-screen"}>
+      <body className={inter.className + " min-h-dvh"}>
         <Providers>
           <div className="flex flex-row justify-start">
             <div>
               <SideBar />
             </div>
-            <div className="h-screen bg-white flex-1 p-4 text-black overflow-auto">
+            <div className="h-dvh bg-white flex-1 text-black overflow-auto">
               {children}
             </div>
           </div>
