+++ frontend/src/helpers/assignment-service/api-wrapper.ts
@@ -41,12 +41,12 @@ const getAssignmentById = async ({
           <p>&nbsp;</p>
           <p><strong class="example">Example 1:</strong></p>
           <pre><strong>Input:</strong> nums = [2,7,11,15], target = 9
-          <strong>Output:</strong> [0,1]
-          <strong>Explanation:</strong> Because nums[0] + nums[1] == 9, we return [0, 1].
+<strong>Output:</strong> [0,1]
+<strong>Explanation:</strong> Because nums[0] + nums[1] == 9, we return [0, 1].
           </pre>
           <p><strong class="example">Example 2:</strong></p>
           <pre><strong>Input:</strong> nums = [3,2,4], target = 6
-          <strong>Output:</strong> [1,2]
+<strong>Output:</strong> [1,2]
           </pre>
           <p>&nbsp;</p>
           <p><strong>Constraints:</strong></p>
